шок я сделала это запускаемым)))))

+++ src/main/java/ru/renett/service/old/RequestValidator.java
@@ -4,13 +4,13 @@ import ru.renett.exceptions.InvalidRequestDataException;
 
 public class RequestValidator implements RequestValidatorInterface{
     @Override
-    public int checkRequestedIdCorrect(String id) throws InvalidRequestDataException {
-        int parsedValue = 0;
+    public Long checkRequestedIdCorrect(String id) throws InvalidRequestDataException {
+        long parsedValue = 0L;
         if (id == null) {
             throw new InvalidRequestDataException("id is null");
         }
         try {
-            parsedValue = Integer.parseInt(id);
+            parsedValue = Long.parseLong(id);
             if (parsedValue < 1) {
                 throw new InvalidRequestDataException("id value cannot be non positive");
             }
