+++ src/main/java/ru/renett/service/old/userService/UserPreferencesInterface.java
@@ -7,6 +7,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public interface UserPreferencesInterface {
-    void saveLastViewedArticleIdCookie(int id, HttpServletResponse response);
+    void saveLastViewedArticleIdCookie(Long id, HttpServletResponse response);
     Cookie getCookieOfLastViewedArticle(HttpServletRequest request) throws InvalidRequestDataException;
 }
