+++ src/main/java/ru/renett/service/old/userService/UserService.java
@@ -3,7 +3,7 @@ package ru.renett.service.old.userService;
 import ru.renett.models.User;
 
 public interface UserService {
-    User getUserById(int id);
+    User getUserById(Long id);
     void editUserData(User user);
     void deleteUser(User user);
     void addNewUser(User user);
